refactor(Validator): Validator 클래스를 이동 및 통합
- Validator 클래스를 calculator.global.validator로 이동
